+++ gateway/tests/test_background.py
@@ -8,7 +8,7 @@ import pytest
 from gateway.config.configure_gateway_logging import configure_gateway_logging
 from gateway.config.get_gateway_settings import GatewaySettings
 from gateway.core.telegram_gateway_background import start_telegram_gateway_background
-from gateway.start_gateway import (
+from gateway.core.telegram_polling_runtime import (
     initialize_telegram_polling_runtime,
     shutdown_telegram_polling_runtime,
 )
