+++ surfaces/interactive_shell/main.py
@@ -88,7 +88,7 @@ def run_repl(
         from gateway.config.configure_gateway_logging import configure_gateway_logging
         from gateway.config.get_gateway_settings import try_load_gateway_settings_for_startup
         from gateway.core.telegram_gateway_background import start_telegram_gateway_background
-        from gateway.start_gateway import (
+        from gateway.core.telegram_polling_runtime import (
             initialize_telegram_polling_runtime,
             shutdown_telegram_polling_runtime,
         )
